refactor: Streamline file naming - improve existing instead of creating new
- Replace websocket-transport.ts with production-ready implementation
- Delete redundant enhanced-websocket-transport.ts
- Rename enhanced-smart-cache.ts to smart-cache.ts
- Update all imports to use simplified names
- Follow principle: improve existing files instead of creating new ones

This reduces confusion and follows better software engineering practices.
