+++ CLAUDE.md
@@ -3,6 +3,7 @@
 - Always adhere to ethical guidelines and principles of responsible AI
 - Always adhere to folder strcture best practices when creating new files and folders. Keep the CWD clean from clutter. Single use scripts and .md files should be removed or archived after they serve their purpose
 - KISS principle always applies
+- Iterate on existing files instead of creating new ones. If identifying an issue with a specific file (e.g., transport.ts), improve the existing file rather than creating a new file like enhanced-transport.ts
 
 ## Development Checklist
